+++ CHANGELOG.md
@@ -11,6 +11,8 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Added
 
+- `relays` command to `ttn-lw-cli`, which enables the management of LoRaWAN relays, including their configuration and uplink forwarding rules.
+
 ### Changed
 
 ### Deprecated